fix: refined chat state wiring and tab visibility
